prints first 98 fibonacci sequence

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,7 +5,7 @@
   *
   * Return: Nothing
   */
-unsigned long long int main(void)
+int main(void)
 {
 	int i = 0;
 	long j = 1, k = 2;
